[Refactor] Switch to new training code.
