+++ setup.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 from setuptools import find_packages, setup
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"
 
 if __name__ == '__main__':
     setup(
